+++ changelog.md
@@ -21,6 +21,7 @@ Codion Change Log
 - DialogBuilder.onBuild() added.
 - CalendarPanel.doubleClicked() added.
 - DefaultCalendarDialogBuilder now accepts the selected date when the day selection panel is double clicked.
+- FilterTable.Export.replaceNewline(String) added, defaults to a single space, DefaultExport now automatically trims strings. 
 ### is.codion.framework.model
 - DefaultForeignKeyConditionModel now links search models to their respective operands.
 ### is.codion.swing.framework.model